Refactored project structure to match requirements, experimented with different optimizers
