+++ recognition/s4433723_AKOA_laterality_perceiver/README.md
@@ -2,6 +2,8 @@
 
 ![img.png](img.png)
 
+Perceiver Transformer Architecture:
+-
 The Perceiver Transformer is a type of transformer network which relies on cross-attention layers
 instead of self-attention layers. This helps reduce the dimensionality of the model such that
 
@@ -11,7 +13,8 @@ The constructed Perceiver model in the current report has the following structur
 
 
 Requirements:
-- Tensorflow version >= 2.4.1
+- 
+Tensorflow version >= 2.4.1
   - i.e. conda install tensorflow-gpu
 - tensorflow-addons i.e. pip install -U tensorflow-addons
 - tensorflow-datasets i.e. pip install tensorflow-datasets
@@ -25,6 +28,12 @@ Training:
 The images in the AKOA dataset are of size _ * _, which are downsampled to 16 * 16
 for the training and validation of the model. 
 
+![input_train_image.png](input_train_image.png)
+
 
 ![test_acc.png](test_acc.png)
 ![Successful_Training_1.png](Successful_Training_1.png)
+
+
+References:
+-
